% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMDX.R
\name{writeMDX}
\alias{writeMDX}
\title{writeMDX}
\usage{
writeMDX(input, config = list("date", "title", "featuredImage"))
}
\arguments{
\item{input}{Input file (.Rmd format).}

\item{config}{\code{list}. Configurable YAML fields to check for. Default is
config = list("date", "title", "featuredImage")}
}
\description{
write and Rmarkdown (.Rmd) file to MDX (.mdx) format.
}
\examples{
\dontrun{
# supposing you have a 'test.Rmd' file:
writeMDX("test.Rmd")
}
}
